Export installation function
